// packages/shared/src/types.ts
export interface User {
  id: number
  email: string
  created_at: string
  updated_at?: string
}
// ...
export interface AuthResponse {
  token: string
  refresh_token: string
  user: User
}